Tasklist: Make drag&drop in the deskbar mode more intuitive
When sorting order is set to Drag&Drop in the deskbar mode, tasklist
plugin uses the x coordinate to determine where to put dragged buttons.
This behaviour is counter-intuitive.

Add check for deskbar mode in xfce_tasklist_button_drag_data_received,
and use the y coordinate in such case.

I suspect, it was forgotten to be updated when the deskbar mode was
added.